fix: rename and export OakQuizResultItem
