--- src/components/organisms/pupil/quiz/InternalQuizResultItem/index.ts
@@ -1,1 +1,0 @@
-export * from "./InternalQuizResultItem";
