+++ src/components/organisms/pupil/quiz/OakQuizResultItem/index.ts
@@ -1,0 +1,1 @@
+export * from "./OakQuizResultItem";
